code/requirements.txt
google-cloud-storage==2.9.0
google-cloud-secret-manager==2.16.1
google-cloud-bigquery==3.11.4
langchain==0.0.305
langchain-openai==0.0.2
pyyaml==6.0.1
functions-framework==3.4.0
gunicorn==20.1.0
openai>=1.6.1,<2.0.0